backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
google-genai==1.73.1
